fix: enable linting for API app
