+++ package.json
@@ -135,9 +135,12 @@
     "@total-typescript/ts-reset": "^0.6.1",
     "@types/amphtml-validator": "^1.0.4",
     "@types/jest": "^29.5.14",
+    "@types/jsdom": "^21.1.7",
     "@types/node": "^20.19.4",
+    "@types/qrcode": "^1.5.5",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
+    "@typescript-eslint/parser": "^8.40.0",
     "chromatic": "^13.0.1",
     "cross-env": "^7.0.3",
     "cross-fetch": "^4.1.0",
@@ -158,8 +161,6 @@
     "playwright": "^1.53.1",
     "plop": "^4.0.1",
     "polished": "^4.3.1",
-    "@types/jsdom": "^21.1.7",
-    "@types/qrcode": "^1.5.5",
     "prettier": "^3.5.3",
     "prettier-plugin-tailwindcss": "^0.6.13",
     "react-server-dom-webpack": "^19.1.0",
