+++ tsconfig.json
@@ -7,6 +7,7 @@
     { "path": "./apps/cms" },
     { "path": "./apps/shop-abc" },
     { "path": "./apps/shop-bcd" },
+    { "path": "./apps/api" },
     { "path": "packages/platform-core" },
     { "path": "./scripts" }
   ]
